Improve frame data loading
Instead of triggering a goToFrame on <Frame>, which caused a navigation
and messed up the browser history, now <Frame> just asks the worker for
the frame data and uses it :)

Also move "selected item" and `caption` states to <Frame>, as it's the only
component that cares about these. No need to have it on the main <App>
and pass these around along with their setters.